pyproject.toml
[tool.poetry]
name = "zerotwo"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {extras = ["voice"], version = "*"}
wavelink = "*"
jishaku = "*"
aiodns = "*"
cchardet = "^*"
uvloop = {version = "*", platform = "linux"}
psutil = "*"

[tool.poetry.dev-dependencies]
flake8 = "*"
flake8-bugbear = "*"
flake8-comprehensions = "*"
pep8-naming = "*"
bandit = "*"
black = {version = "*", allow-prereleases = true}
pre-commit = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
